some change

+++ modules/task_1/golubeva_a_mismatched_items/mismatched_items.cpp
@@ -70,7 +70,7 @@ int getCountDiffChar(char *str1, char *str2) {
 
 int getCount(char* str1, char* str2) {
   int counter = 0;
-  for (int i = 0; i < strlen(str1); i++) {
+  for (size_t i = 0; i < strlen(str1); i++) {
     if (str1[i] != str2[i])
       counter++;
   }
